refactor(place): bottom sheet 관련 리팩토링
- 추후 여행도 보여주기 위해 place_sheet를 bottomSheetWidget으로 변경
- PlaceSimpleView, PlaceListView를 각각 다른 파일로 분리
