+++ lib/ui/main_page.dart
@@ -5,7 +5,7 @@ import 'package:yeohaeng_ttukttak/states/navigation_state.dart';
 import 'package:yeohaeng_ttukttak/states/place_view_model.dart';
 import 'package:yeohaeng_ttukttak/ui/main/custom_google_map/my_location_button_widget.dart';
 import 'package:yeohaeng_ttukttak/ui/place/place_list_view.dart';
-import 'package:yeohaeng_ttukttak/ui/place/place_sheet_widget.dart';
+import 'package:yeohaeng_ttukttak/ui/place/bottom_sheet_widget.dart';
 import 'package:yeohaeng_ttukttak/ui/place/place_type_filter_widget.dart';
 import 'package:yeohaeng_ttukttak/ui/main/custom_google_map/map_search_bar.dart';
 import 'package:yeohaeng_ttukttak/ui/main/custom_google_map/search_nearby_button_widget.dart';
@@ -89,7 +89,7 @@ class _MainPageState extends State<MainPage> with TickerProviderStateMixin {
           const MyLocationButtonWidget(),
           if (isSheetShown)
             Positioned(
-                bottom: 0.0, left: 0.0, right: 0.0, child: PlaceViewSheet())
+                bottom: 0.0, left: 0.0, right: 0.0, child: BottomSheetWidget())
         ]);
       }),
       floatingActionButton: isSheetShown && isPlaceSelected
